+++ src/ambition/version_details.cc
@@ -46,6 +46,7 @@ std::vector<VersionUpdates> getVersionDetails(
         .items = {
           _("Idle caravans now show a message in place of the clone button."),
           _("Caravans now show as idle (with an asterisk) on the Trade Report when they are considered idle by the automatic caravan replacement and cloning system."),
+          _("Draw building link icons at a constant FPS, independent of game speed."),
           _("Draw ship washes (the waves around, and the wake behind, ships) at a constant FPS, independent of game speed."),
           _("Do not show ship washes for ships that are stopped and not moving."),
           _("Made development slightly easier."),